Moved Screen Orientation to Shared
