+++ src/pages/journal/__tests__/journal.spec.ts
@@ -19,7 +19,7 @@ import { Subscription } from 'rxjs/Subscription';
 import { SlotSelectorProvider } from '../../../providers/slot-selector/slot-selector';
 import { MockedJournalModule } from '../__mocks__/journal.module.mock';
 import { ScreenOrientation } from '@ionic-native/screen-orientation';
-import { ScreenOrientationMock } from '../components/test-slot/__mocks__/screen-orientation.mock';
+import { ScreenOrientationMock } from '../../../shared/mocks/screen-orientation.mock';
 import { UnloadJournal, LoadJournal, LoadJournalSuccess } from '../journal.actions';
 import { BasePageComponent } from '../../../shared/classes/base-page';
 import { StoreModel } from '../../../shared/models/store.model';
