+++ src/shared/mocks/screen-orientation.mock.ts
